+++ components/philips_series_2200/switch/power.h
@@ -98,7 +98,7 @@ namespace esphome
                 /// @brief Determines wether a power trip should be performed
                 bool should_power_trip_ = false;
                 /// @brief Time of last power trip
-                long last_power_trip_ = 0;
+                uint32_t last_power_trip_ = 0;
                 /// @brief nr of power performed power trips
                 int power_trip_count_ = 0;
                 /// @brief initial power state reference
